notification: Fix duplicate notifications
GNOME Shell uses a proxy to listen to the org.freedesktop.Notifications
well-known name. This is for sandboxing: access to send notifications doesn't
give access to the shell's other D-Bus APIs. The proxy resends all
notifications to the shell.

This results in GSConnect receiving duplicate notifications, since it listens
for all calls to Notify on the session bus. It would receive both the original
notification and the resent one.

Fixed by ignoring notifications sent from the owner of
org.freedesktop.Notifications. This involves subscribing to NameOwnerChanged
in case the notification proxy is restarted.